+++ src/Core/ViewConfig.php
@@ -16,31 +16,6 @@ use OxidEsales\GdprOptinModule\Transput\OptInRequestInterface;
  */
 class ViewConfig extends ViewConfig_parent
 {
-    public function showGdprInvoiceOptIn(): bool
-    {
-        return $this->getService(ModuleSettingsInterface::class)->showInvoiceOptIn();
-    }
-
-    public function showGdprDeliveryOptIn(): bool
-    {
-        return $this->getService(ModuleSettingsInterface::class)->showDeliveryOptIn();
-    }
-
-    public function showGdprRegistrationOptIn(): bool
-    {
-        return $this->getService(ModuleSettingsInterface::class)->showRegistrationOptIn();
-    }
-
-    public function showGdprReviewOptIn(): bool
-    {
-        return $this->getService(ModuleSettingsInterface::class)->showReviewOptIn();
-    }
-
-    public function getGdprContactOptIn(): string
-    {
-        return $this->getService(ModuleSettingsInterface::class)->getContactOptIn();
-    }
-
     public function getInvoiceOptIn(): bool
     {
         return $this->getService(OptInRequestInterface::class)->getInvoiceAddressOptIn();
@@ -49,5 +24,10 @@ class ViewConfig extends ViewConfig_parent
     public function getDeliveryOptIn(): bool
     {
         return $this->getService(OptInRequestInterface::class)->getDeliveryAddressOptIn();
+    }
+
+    public function getGdprModuleSettings(): ModuleSettingsInterface
+    {
+        return $this->getService(ModuleSettingsInterface::class);
     }
 }