+++ tests/Integration/Core/ViewConfigTest.php
@@ -13,99 +13,23 @@ use OxidEsales\Eshop\Core\ViewConfig;
 use OxidEsales\EshopCommunity\Internal\Framework\Module\Facade\ModuleSettingServiceInterface;
 use OxidEsales\GdprOptinModule\Core\GdprOptinModule;
 use OxidEsales\GdprOptinModule\Service\ModuleSettings;
+use OxidEsales\GdprOptinModule\Service\ModuleSettingsInterface;
 use OxidEsales\GdprOptinModule\Tests\Integration\BaseTestCase;
 use OxidEsales\GdprOptinModule\Tests\Traits\ServiceContainer;
 use OxidEsales\GdprOptinModule\Transput\OptInRequestInterface;
 
 final class ViewConfigTest extends BaseTestCase
 {
-    use ServiceContainer;
-
-    public static function dataProviderBoolean(): array
+    public function testGetGdprModuleSettings(): void
     {
-        return [
-            'on' => [true],
-            'off' => [false],
-        ];
-    }
+        $moduleSettingsStub = $this->createStub(ModuleSettingsInterface::class);
 
-    public function testGetGdprContactOptIn(): void
-    {
-        $value = 'the_value';
+        $sut = $this->createPartialMock(\OxidEsales\GdprOptinModule\Core\ViewConfig::class, ['getService']);
+        $sut->method('getService')->willReturnMap([
+            [ModuleSettingsInterface::class, $moduleSettingsStub]
+        ]);
 
-        $settingsService = $this->getServiceFromContainer(ModuleSettingServiceInterface::class);
-        $settingsService->saveString(
-            ModuleSettings::CONTACT_CHOICE,
-            $value,
-            GdprOptinModule::MODULE_ID
-        );
-
-        $viewConfig = oxNew(ViewConfig::class);
-        $this->assertSame($value, $viewConfig->getGdprContactOptIn());
-    }
-
-    /**
-     * @dataProvider dataProviderBoolean
-     */
-    public function testShowGdprInvoiceOptIn(bool $value): void
-    {
-        $settingsService = $this->getServiceFromContainer(ModuleSettingServiceInterface::class);
-        $settingsService->saveBoolean(
-            ModuleSettings::INVOICE_OPT_IN,
-            $value,
-            GdprOptinModule::MODULE_ID
-        );
-
-        $viewConfig = oxNew(ViewConfig::class);
-        $this->assertSame($value, $viewConfig->showGdprInvoiceOptIn());
-    }
-
-    /**
-     * @dataProvider dataProviderBoolean
-     */
-    public function testShowGdprDeliveryOptIn(bool $value): void
-    {
-        $settingsService = $this->getServiceFromContainer(ModuleSettingServiceInterface::class);
-        $settingsService->saveBoolean(
-            ModuleSettings::DELIVERY_OPT_IN,
-            $value,
-            GdprOptinModule::MODULE_ID
-        );
-
-        $viewConfig = oxNew(ViewConfig::class);
-        $this->assertSame($value, $viewConfig->showGdprDeliveryOptIn());
-    }
-
-    /**
-     * @dataProvider dataProviderBoolean
-     */
-    public function testShowGdprRegistrationOptIn(bool $value): void
-    {
-        $settingsService = $this->getServiceFromContainer(ModuleSettingServiceInterface::class);
-        $settingsService->saveBoolean(
-            ModuleSettings::REGISTRATION_OPT_IN,
-            $value,
-            GdprOptinModule::MODULE_ID
-        );
-
-        $viewConfig = oxNew(ViewConfig::class);
-        $this->assertSame($value, $viewConfig->showGdprRegistrationOptIn());
-    }
-
-    /**
-     * @dataProvider dataProviderBoolean
-     */
-    public function testShowGdprReviewOptIn(bool $value): void
-    {
-        $settingsService = $this->getServiceFromContainer(ModuleSettingServiceInterface::class);
-        $settingsService->saveBoolean(
-            ModuleSettings::REVIEW_OPT_IN,
-            $value,
-            GdprOptinModule::MODULE_ID
-        );
-
-        $viewConfig = oxNew(ViewConfig::class);
-        $this->assertSame($value, $viewConfig->showGdprReviewOptIn());
+        $this->assertSame($moduleSettingsStub, $sut->getGdprModuleSettings());
     }
 
     public function testGetInvoiceOptIn(): void